requirements.txt
flet>=0.22.0
openai>=1.30.0
python-dotenv>=1.0.1
mysql-connector-python>=9.0.0
numpy>=1.26.4
beautifulsoup4>=4.12.3
requests>=2.32.3
